wip: Adding new menu options to do stuff
