+++ pyproject.toml
@@ -32,7 +32,7 @@ classifiers = [
 
 keywords = ["game", "engine", "dice"]
 
-dependencies = ["pydantic>=2.9.2", "click>=8.1.7"]
+dependencies = ["pydantic>=2.9.2", "click>=8.1.7", "pyyaml>=6.0"]
 
 [project.urls]
 Homepage = "https://github.com/carrera-dev-consulting/zombie_nomnom"
